Extract string utils to separate file
